simplify dependencies, remove some cycles
