+++ components/fdk-aac_decoder/fdk_aac_decoder.c
@@ -16,8 +16,9 @@
 #include "driver/i2s.h"
 
 #include "common_buffer.h"
+#include "common_component.h"
 #include "aacdecoder_lib.h"
-#include "audio_player.h"
+#include "audio_renderer.h"
 #include "m4a.h"
 
 #define TAG "fdkaac_decoder"
