+++ components/mp3_decoder/mp3_decoder.c
@@ -20,7 +20,6 @@
 
 #include "driver/i2s.h"
 #include "audio_renderer.h"
-#include "audio_player.h"
 #include "spiram_fifo.h"
 #include "mp3_decoder.h"
 #include "common_buffer.h"
